# doc/openapi/modlink-platform.openapi.yaml
openapi: 3.1.0
info:
  title: ModLinkCloud Platform API
  version: "1.4.0"
  description: |
    控制台与管理后台使用的业务 API。与人读文档：`doc/模链云-API接口文档.md`。

    - **命名空间**：统一前缀 **`/mlk`**（共机部署时便于 `location /mlk/` 分流）。
    - **Base**：`https://api.{domain}/mlk/platform/v1`
    - **鉴权**：`Authorization: Bearer <JWT>`；租户上下文可选 **`X-Org-Id`**（优先于 JWT 默认 org）
    - **响应包裹**：`{ code, message, data, request_id }`（`code=0` 成功）
servers:
  - url: https://api.{domain}
    description: 生产（将 {domain} 换为实际备案域名）
security:
  - bearerJwt: []
...
paths:
...
  /mlk/platform/v1/usage/summary:
    get:
      tags: [Usage]
      operationId: usageSummary
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - name: org_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XOrgId:
      name: X-Org-Id
      in: header
      required: false
      schema:
        type: string
      description: 指定当前组织（合法成员时优先于 JWT 默认 org）
...
  schemas:
    ApiEnvelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
        data: {}
        request_id:
          type: string
...